requirements.txt
flask
line-bot-sdk
google-generativeai>=0.3.2
google-auth         # ✅ 新增這行
gunicorn
Pillow
python-dotenv